+++ src/impl/gameplay.c
@@ -16,6 +16,8 @@
 #include "../ent/player.h"
 #include "../ent/rack.h"
 #include "../str/rack_string.h"
+#include "../util/io_util.h"
+#include "../util/string_util.h"
 #include "move_gen.h"
 #include <assert.h>
 #include <stdbool.h>
@@ -207,8 +209,8 @@ bool draw_rack_from_bag(const Game *game, const int player_index,
   const uint16_t dist_size = rack_get_dist_size(player_rack);
   rack_copy(player_rack, rack_to_draw);
   for (int i = 0; i < dist_size; i++) {
-    const int8_t rack_number_of_letter = rack_get_letter(player_rack, i);
-    for (int8_t j = 0; j < rack_number_of_letter; j++) {
+    const uint16_t rack_number_of_letter = rack_get_letter(player_rack, i);
+    for (uint16_t j = 0; j < rack_number_of_letter; j++) {
       if (!bag_draw_letter(bag, i, player_draw_index)) {
         return false;
       }
@@ -224,8 +226,8 @@ void draw_leave_from_bag(Bag *bag, int player_draw_index, Rack *rack_to_update,
                          const Rack *rack_to_draw) {
   const uint16_t dist_size = rack_get_dist_size(rack_to_draw);
   for (int i = 0; i < dist_size; i++) {
-    const int8_t rack_number_of_letter = rack_get_letter(rack_to_draw, i);
-    for (int8_t j = 0; j < rack_number_of_letter; j++) {
+    const uint16_t rack_number_of_letter = rack_get_letter(rack_to_draw, i);
+    for (uint16_t j = 0; j < rack_number_of_letter; j++) {
       if (!bag_draw_letter(bag, i, player_draw_index)) {
         continue;
       }
@@ -265,8 +267,8 @@ void return_rack_to_bag(const Game *game, const int player_index) {
   int player_draw_index = game_get_player_draw_index(game, player_index);
   const uint16_t dist_size = rack_get_dist_size(player_rack);
   for (int i = 0; i < dist_size; i++) {
-    const int8_t rack_number_of_letter = rack_get_letter(player_rack, i);
-    for (int8_t j = 0; j < rack_number_of_letter; j++) {
+    const uint16_t rack_number_of_letter = rack_get_letter(player_rack, i);
+    for (uint16_t j = 0; j < rack_number_of_letter; j++) {
       bag_add_letter(bag, i, player_draw_index);
     }
   }
